tools/tts/python/xtts/pyproject.toml
[project]
name = "lecture-tts-xtts"
version = "0.1.0"
description = "XTTS-v2 (Coqui) TTS provider for lecture-automation. Coqui Public Model License."
requires-python = ">=3.10,<3.13"
dependencies = [
    # 메인테인드 community fork. 원본 idiap/coqui-ai-TTS.
    # 모델 가중치는 별도 라이선스 (Coqui Public Model License).
    "coqui-tts>=0.24.0",
    "numpy>=1.24,<2.0",
    "torch>=2.0,<2.5",
    "torchaudio>=2.0,<2.5",
]
